requirements.txt
requests==2.31.0
python-binance==1.0.19
openai>=1.30.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
ta==0.11.0
flask==3.0.0
flask-cors==4.0.0

